refactor: move export related api call to its own file
